requirements.txt
numpy==2.0.2
pillow==11.3.0